Generalize BenchmarkRunner (#572)
This PR generalizes `BenchmarkRunner`, `BenchmarkConfig`, `BenchmarkResult`, and `Benchmark` to `AbstractBenchmarkRunner`, `AbstractBenchmarkConfig`, `AbstractBenchmarkResult`, and `AbstractBenchmark`. Thereby, the former classes become `LambdaBenchmarkRunner`, `LambdaBenchmarkConfig`, `LambdaBenchmarkResult` and `LambdaBenchmark`. 

The derived classes still have plenty logic that might go in the abstract ones in the future. But for the sake of keeping this PR as small as possible, I suggest to postpone that to the point where we introduce another set of benchmark classes like `EC2Benchmark*`