Don't fail if Skyrise is built outside of a git repository (#658)
The latest changes in our third party dependency cmake-git-version-tracking makes the build process fail, if not .git folder is found. This affects all build setups, where the git directory is not inside the working directory (e.g. by downloading the source code as zip file from github).

This PR restores the old behaviour. An empty string is written as commit id if not repository is found.

Co-authored-by: Thomas Bodner <thomas.bodner@hpi.de>