Improved Stream handling when receiving data from S3 (#721)
This PR changes the storage interface. To avoid copying ObjectReader::read now takes a pointer to a std::vector<char> as destination. Because this vector can transparently be cleared this allows us to fix a previous error, where we mixed up actual data and error messages from the SDK. Also, since in almost all use-cases the data from ::read is written to an vector, this simplifies using the storage classes. Tests are adapted accordingly.

Co-authored-by: Thomas Bodner <thomas.bodner@hpi.de>