ORC Reader Caching (#778)
This PR introduces the BufferedReader and integrates it with the OrcFormatReader. A BufferedReader can be constructed from any ObjectReader and is itself an ObjectReader. It intercepts calls to Read() and ReadTail() and is meant to either query and cache larger ranges from the underlying source, or serve the calls from its internal cache. The caching behavior is controlled by an CachableLocations class. It marks certain areas of the source object as cachable. These ranges are represented by the class BufferLocation.

The caching algorithm receives the list of cachable locations and a desired cache size as input. For each query that cannot be entirely served from the cache, the cacheable locations with the minimum difference in size to the desired cache size that includes the queries bytes is determined and received from the source. Cacheable locations larger than the desired cache size are adjusted to exactly fit the desired size.

Co-authored-by: Thomas Bodner <thomas.bodner@hpi.de>